+++ src/Service/BatchTutoringSessionCreationService.php
@@ -2,10 +2,11 @@
 
 namespace App\Service;
 
-use App\Constants;
+use App\Entity\Student;
 use App\Entity\TutoringSession;
 use App\Exception\BatchTutoringSessionCreationException;
 use App\Model\BatchTutoringSessionCreationModel;
+use App\Utils\DateUtils;
 use DateTimeInterface;
 use Doctrine\ORM\EntityManagerInterface;
 use Exception;
@@ -23,18 +24,19 @@ readonly class BatchTutoringSessionCreationService
 
     public function batchCreateSessions(BatchTutoringSessionCreationModel $batchTutoringSessionCreationModel): void
     {
+        /** @var Student $user */
         $user = $this->security->getUser();
 
         $selectedWeekdays = $batchTutoringSessionCreationModel->getWeekDays();
         /** @var DateTimeInterface[] $dates */
-        $dates = $this->getAllDatesBetweenDatesByWeekdays($batchTutoringSessionCreationModel->getStartDate(), $batchTutoringSessionCreationModel->getEndDate(), $selectedWeekdays);
+        $dates = DateUtils::getAllDatesBetweenDatesByWeekdays($batchTutoringSessionCreationModel->getStartDate(), $batchTutoringSessionCreationModel->getEndDate(), $selectedWeekdays);
         $this->entityManager->getConnection()->beginTransaction();
         try {
             foreach ($dates as $date) {
                 $startDateTime = clone $date->setTime(hour: $batchTutoringSessionCreationModel->getStartTime()->format('H'), minute: $batchTutoringSessionCreationModel->getStartTime()->format('i'));
                 $endDateTime = clone $date->setTime(hour: $batchTutoringSessionCreationModel->getEndTime()->format('H'), minute: $batchTutoringSessionCreationModel->getEndTime()->format('i'));
 
-                $turoringSession = (new TutoringSession())
+                $tutoringSession = (new TutoringSession())
                     ->setBuilding($batchTutoringSessionCreationModel->getBuilding())
                     ->setRoom($batchTutoringSessionCreationModel->getRoom())
                     ->setCreatedBy($user)
@@ -43,7 +45,7 @@ readonly class BatchTutoringSessionCreationService
                     ->setEndDateTime($endDateTime)
                     ->addTutor($user)
                 ;
-                $this->entityManager->persist($turoringSession);
+                $this->entityManager->persist($tutoringSession);
             }
             $this->entityManager->flush();
             $this->entityManager->getConnection()->commit();
@@ -55,35 +57,5 @@ readonly class BatchTutoringSessionCreationService
 
             throw new BatchTutoringSessionCreationException($e->getMessage(), $e->getCode(), $e);
         }
-    }
-
-    private function getAllDatesBetweenDatesByWeekdays(DateTimeInterface $startDate, DateTimeInterface $endDate, array $selectedWeekdays): array
-    {
-        $result = [];
-        $selectedWeekdaysIndexes = [];
-        $weekDays = Constants::getAvailableWeekdays();
-
-        foreach ($weekDays as $index => $dayName) {
-            if (in_array($dayName, $selectedWeekdays)) {
-                $selectedWeekdaysIndexes[] = $index + 1;
-            }
-        }
-
-        $currentDate = clone $startDate;
-        while ($currentDate <= $endDate) {
-            $dayOfWeek = $currentDate->format('N');
-
-            if (in_array($dayOfWeek, $selectedWeekdaysIndexes)) {
-                $result[] = clone $currentDate;
-            }
-            if ($dayOfWeek > 4) {
-                // Skip Saturday and Sunday
-                $currentDate->modify(sprintf('+%s day', 8 - $dayOfWeek));
-            } else {
-                $currentDate->modify('+1 day');
-            }
-        }
-
-        return $result;
     }
 }