+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - Upgrade to symfony v6.3 !29
 
+## Fixes
+- Add `AllDefaultWeekdaysHaveAtLeastOneSession` constraint + fix all validation wording !28
+
 ## [1.2.0] - 2023-09-25
 
 ## Added